Implemented Linked list + additional functionalities.
Created Linked List data structure with a prev and a next reference in
each node. The nodes contain an Image as their data object.

Created MyListIterator interface(Generic) and implemented it in the
Linked List so that you can use this iterator to go forward, back, and
start back at the beginning. 

Added forward button functionality (to go through the next images) and
it can be called forever essentially, since it restarts to the beginning
of the list.

Made use of the accordion(Titled pane) to show the user the list of
images that were retrieved from the folder.